build out for roles config L2 input file validation
